Extend ESLint (sorting imports, removing file extensions). Get files structure more organized.
