+++ src/api/constants.ts
@@ -1,0 +1,2 @@
+export const MOVIE_API_BASE_URL = 'https://api.themoviedb.org/3';
+export const IMAGE_BASE_URL = 'https://image.tmdb.org/t/p/w500';
