+++ src/api/fetchMovieDetails.ts
@@ -1,5 +1,5 @@
-import type {Movie} from './types.ts';
-import {movieApiService} from './movieApiService.ts';
+import type {Movie} from './types';
+import {movieApiService} from './movieApiService';
 
 export const fetchMovieDetails = async (movieId: string): Promise<Movie> => {
   return (await movieApiService.get(`/movie/${movieId}`)).data;
