+++ src/api/movieApiService.ts
@@ -1,7 +1,8 @@
 import axios from 'axios';
+import {MOVIE_API_BASE_URL} from './constants';
 
 export const movieApiService = axios.create({
-  baseURL: 'https://api.themoviedb.org/3',
+  baseURL: MOVIE_API_BASE_URL,
   params: {
     api_key: import.meta.env.VITE_MOVIES_DATABASE_API_KEY,
   },
