+++ src/main.tsx
@@ -1,6 +1,6 @@
 import {StrictMode} from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './App.tsx';
+import App from './app/App';
 import './styles/global.css';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
