+++ src/pages/MoviesListPage/components/EmptyInitialState/icons/EmtpyStateIcon.tsx
@@ -1,4 +1,4 @@
-import {Icon} from '../EmptyInitialState.styled.tsx';
+import {Icon} from '../EmptyInitialState.styled';
 
 export default function EmptyStateIcon() {
   return (
