+++ src/pages/MoviesListPage/components/EmptySearchResult/NothingFoundState.tsx
@@ -1,5 +1,5 @@
-import {Container, Content, Message, Title} from './NothingFoundState.styled.tsx';
-import NothingFoundIcon from './icons/NothingFoundIcon.tsx';
+import {Container, Content, Message, Title} from './NothingFoundState.styled';
+import NothingFoundIcon from './icons/NothingFoundIcon';
 
 export default function NothingFoundState() {
   return (
