+++ src/routes/AppRoutes.tsx
@@ -1,9 +1,9 @@
 import {lazy} from 'react';
 import {Route, Routes} from 'react-router-dom';
-import {PageType, PathByPageType} from './constants.ts';
+import {PageType, PathByPageType} from './constants';
 
-const MoviesListPage = lazy(() => import('../pages/MoviesListPage/MoviesListPage.tsx'));
-const MovieDetailsPage = lazy(() => import('../pages/MovieDetailsPage/MovieDetailsPage.tsx'));
+const MoviesListPage = lazy(() => import('../pages/MoviesListPage/MoviesListPage'));
+const MovieDetailsPage = lazy(() => import('../pages/MovieDetailsPage/MovieDetailsPage'));
 
 export default function AppRoutes() {
   return (
