+++ src/types/styles.d.ts
